refactor _openai module for better naming
add new file config.py for templating
